refactor: 카드 생성자의 접근제어자 package private 으로 변경
