+++ src/main/java/blakjack/domain/card/Card.java
@@ -4,7 +4,7 @@ public final class Card {
     private final Suit suit;
     private final Face face;
 
-    public Card(final Suit suit, final Face face) {
+    Card(final Suit suit, final Face face) {
         this.suit = suit;
         this.face = face;
     }
