+++ src/test/java/blakjack/domain/card/MockCard.java
@@ -1,10 +1,6 @@
-package blakjack.domain;
+package blakjack.domain.card;
 
-import blakjack.domain.card.Card;
-import blakjack.domain.card.Face;
-import blakjack.domain.card.Suit;
-
-public class Fixture {
+public final class MockCard {
     public final static Card CLUB_TEN = new Card(Suit.CLUB, Face.TEN);
     public final static Card HEART_ACE = new Card(Suit.HEART, Face.ACE);
     public final static Card CLUB_NINE = new Card(Suit.CLUB, Face.NINE);
