+++ src/test/java/blakjack/domain/state/gameResult/WinTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static blakjack.domain.Fixture.CLUB_TEN;
+import static blakjack.domain.card.MockCard.CLUB_TEN;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class WinTest {
